pyproject.toml
[project]
name = "techchallenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcus",email = "marcus_es21@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "flasgger (>=0.9.7.1,<0.10.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "selenium (>=4.34.2,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]

[tool.poetry]
packages = [{include = "techchallenge", from = "src"}]
package-mode=false



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
